FIX:I arranged the folders and files according to the architecture and separated the types and their responsibilities
